+++ tasks/smyshlaev_a_str_order_check/seq/src/ops_seq.cpp
@@ -28,9 +28,9 @@ bool SmyshlaevAStrOrderCheckSEQ::RunImpl() {
     const std::string& str1 = input_data.first;
     const std::string& str2 = input_data.second;
 
-    size_t min_len = std::min(str1.length(), str2.length());
+    int min_len = std::min(str1.length(), str2.length());
 
-    for (size_t i = 0; i < min_len; ++i) {
+    for (int i = 0; i < min_len; ++i) {
         if (str1[i] < str2[i]) {
             GetOutput() = -1;
             return true;
